alerting: add labels to name when converting to series also move conversion func to alerting package since it is not used in other places related to #28068
