+++ pkg/tsdb/frame_util.go
@@ -1,12 +1,9 @@
 package tsdb
 
 import (
-	"fmt"
 	"time"
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/grafana/grafana/pkg/components/null"
-	"github.com/grafana/grafana/pkg/util/errutil"
 )
 
 // SeriesToFrame converts a TimeSeries to a sdk Frame
@@ -37,58 +34,4 @@ func convertTSDBTimePoint(point TimePoint) (t *time.Time, f *float64) {
 		f = &point[valueIdx].Float64
 	}
 	return
-}
-
-// FrameToSeriesSlice converts a frame that is a valid time series as per data.TimeSeriesSchema()
-// to a TimeSeriesSlice.
-func FrameToSeriesSlice(frame *data.Frame) (TimeSeriesSlice, error) {
-	tsSchema := frame.TimeSeriesSchema()
-	if tsSchema.Type == data.TimeSeriesTypeNot {
-		// If no fields, or only a time field, create an empty TimeSeriesSlice with a single
-		// time series in order to trigger "no data" in alerting.
-		if len(frame.Fields) == 0 || (len(frame.Fields) == 1 && frame.Fields[0].Type().Time()) {
-			return TimeSeriesSlice{{
-				Name:   frame.Name,
-				Points: make(TimeSeriesPoints, 0),
-			}}, nil
-		}
-		return nil, fmt.Errorf("input frame is not recognized as a time series")
-	}
-
-	seriesCount := len(tsSchema.ValueIndices)
-	seriesSlice := make(TimeSeriesSlice, 0, seriesCount)
-	timeField := frame.Fields[tsSchema.TimeIndex]
-	timeNullFloatSlice := make([]null.Float, timeField.Len())
-
-	for i := 0; i < timeField.Len(); i++ { // built slice of time as epoch ms in null floats
-		tStamp, err := timeField.FloatAt(i)
-		if err != nil {
-			return nil, err
-		}
-		timeNullFloatSlice[i] = null.FloatFrom(tStamp)
-	}
-
-	for _, fieldIdx := range tsSchema.ValueIndices { // create a TimeSeries for each value Field
-		field := frame.Fields[fieldIdx]
-		ts := &TimeSeries{
-			Name:   field.Name,
-			Tags:   field.Labels.Copy(),
-			Points: make(TimeSeriesPoints, field.Len()),
-		}
-
-		for rowIdx := 0; rowIdx < field.Len(); rowIdx++ { // for each value in the field, make a TimePoint
-			val, err := field.FloatAt(rowIdx)
-			if err != nil {
-				return nil, errutil.Wrapf(err, "failed to convert frame to tsdb.series, can not convert value %v to float", field.At(rowIdx))
-			}
-			ts.Points[rowIdx] = TimePoint{
-				null.FloatFrom(val),
-				timeNullFloatSlice[rowIdx],
-			}
-		}
-
-		seriesSlice = append(seriesSlice, ts)
-	}
-
-	return seriesSlice, nil
 }
